fix shell word handling
hss is incompatible with ssh with regard to handling commandlines like:

  hss somehost grep "Some Pattern" /var/log/messages

Although hss tries to quote compound words, using the default
shell-executing behavior of Ruby's exec() causes the quotes to be
consumed. This can be particularly frustrating with a command like grep
which defaults to stdin for input, as the command will hang
indefinitely.

In this change, we use the non-shell variant of exec(), which eliminates
the need for quoting grouped words.